Factoring some related test cases into a new ExecutorStepTest.
